requirements.txt
bidict==0.22.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.4
cycler==0.11.0
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-SocketIO==5.2.0
Flask-WTF==1.0.1
fonttools==4.34.4
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
interchange==2021.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.2
mongoengine==0.24.1
monotonic==1.6
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
pansi==2020.7.3
Pillow==9.2.0
py2neo==2021.2.3
Pygments==2.13.0
pymongo==3.5.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-engineio==4.3.3
python-socketio==5.7.1
pytz==2022.2.1
requests==2.28.1
scipy==1.9.1
six==1.16.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.1.2
WTForms==3.0.1
xlrd2==1.3.4
zipp==3.8.0
